auth-service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Michael <mihail.tchernov@yandex.ru>"]
readme = "README.md"
packages = [{include = "auth_service"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.2"
pydantic = "^1.10.8"
uvicorn = "^0.22.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.15"}
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
